appointment

+++ php/DeleteAppointment.php
@@ -1,6 +1,11 @@
 <?php
 include 'db_connection.php';
 
+if(!$_SESSION['logado']) {
+    header("Location: Login.php");
+    exit;
+}
+
 $id = $_POST['id'];
 $query = $conn->prepare("DELETE FROM consultas WHERE id = :id");
 $query->bindValue(':id', $id);
